Move implementation to `execute.rs`
